requirements.txt
Flask==3.0.2
Flask_QRcode==3.1.0
opencv_python==4.8.0.74
Pillow==10.0.0
Pillow==10.2.0
pytesseract==0.3.10
python-dotenv==1.0.1
pyzbar==0.1.9
Werkzeug==3.0.1
